added a mock backend in python to test the api and updated symfony logic
Python mock backend is added that sends a request to the api. This can be tested by pulling the repo and going to the python plots api folder on your terminal and running docker compose up. That will generate the plot in the mockbackend folder. For the rest updated some logic on the symfony server. Made a succesful connection to python api but havnt been able to show the image on a route.